Move Enums somewhere else to improve API documentation
