+++ darts/__init__.py
@@ -5,27 +5,7 @@ darts
 
 from .timeseries import TimeSeries
 import matplotlib as mpl
-import matplotlib.pyplot as plt
 from matplotlib import cycler
-
-# Enums
-from enum import Enum
-
-
-class SeasonalityMode(Enum):
-    MULTIPLICATIVE = 'multiplicative'
-    ADDITIVE = 'additive'
-    NONE = None
-
-
-class TrendMode(Enum):
-    LINEAR = 'linear'
-    EXPONENTIAL = 'exponential'
-
-
-class ModelMode(Enum):
-    MULTIPLICATIVE = 'multiplicative'
-    ADDITIVE = 'additive'
 
 
 __version__ = 'dev'
