+++ darts/models/ensemble_model.py
@@ -1,6 +1,6 @@
 """
-Ensemble model
---------------
+Ensemble Model Base Class
+-------------------------
 """
 
 from abc import abstractmethod
