+++ darts/utils/statistics.py
@@ -16,7 +16,7 @@ from statsmodels.tsa.stattools import acf
 from ..logging import raise_log, get_logger, raise_if_not
 from ..timeseries import TimeSeries
 from .missing_values import fill_missing_values
-from .. import SeasonalityMode, ModelMode
+from .utils import SeasonalityMode, ModelMode
 
 logger = get_logger(__name__)
